requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cytoolz==0.12.2
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
eth-abi==4.2.1
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.1
exceptiongroup==1.1.3
fastapi==0.103.1
flake8==6.1.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
hexbytes==0.3.1
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
orjson==3.9.7
parsimonious==0.9.0
protobuf==4.24.3
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycodestyle==2.11.0
pycryptodome==3.19.0
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyflakes==3.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pyunormalize==15.0.0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.3
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
toolz==0.12.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
web3==6.9.0
websockets==11.0.3
yarl==1.9.2
